requirements.txt
flask
twilio
python-dotenv
wave
PyAudioWPatch